Move package "client/promise" to "core/promises"
